+++ internal/server/chain_builder.go
@@ -5,6 +5,7 @@ import (
 	"github.com/tingly-dev/tingly-box/internal/obs"
 	"github.com/tingly-dev/tingly-box/internal/protocol"
 	"github.com/tingly-dev/tingly-box/internal/protocol/transform"
+	servertransform "github.com/tingly-dev/tingly-box/internal/server/transform"
 	"github.com/tingly-dev/tingly-box/internal/typ"
 )
 
@@ -29,8 +30,8 @@ func (s *Server) BuildTransformChain(c *gin.Context, targetType protocol.APIType
 	// 2. Base transform (protocol conversion)
 	transforms = append(transforms, transform.NewBaseTransform(targetType))
 	if s.mcpEnabled() {
-		transforms = append(transforms, NewMCPToolInjectionTransform(s))
-		transforms = append(transforms, NewMCPToolStripGuardTransform(s))
+		transforms = append(transforms, servertransform.NewMCPToolInjectionTransform(s.mcpRuntime))
+		transforms = append(transforms, servertransform.NewMCPToolStripGuardTransform(s.mcpRuntime, s.mcpStripDisabledToolsEnabled()))
 	}
 	transforms = append(transforms, transform.NewVendorTransform(providerURL))
