+++ internal/server/transform/mcp_merge_tools.go
@@ -1,4 +1,4 @@
-package server
+package transform
 
 import (
 	"github.com/anthropics/anthropic-sdk-go"
