+++ internal/server/transform/transform_mcp_tool_strip_guard_test.go
@@ -1,4 +1,4 @@
-package server
+package transform
 
 import (
 	"encoding/json"
